Making BingSquad evaluation a module; mainly for use in testing. (#9)
* simplifying bert config files

* removed test argument

* removed test argument flag

* applying comment
